Add a bg image, odin logo and mostly of html
Just realized I wrote almost the entire html file
without commiting.
I downloaded a bg image from unsplash, the odin
logo, wrote the left header and mostly of the
main container. Tried to separate things neatly
with divs, each in their own containers.
Perhaps too many classes? Not quite sure if the
spans are necessary either, but that's up for the
css phase of the project.
Decided to separate the middle container into left
and right, columns instead of rows.