Organize images in folder and position header
I created a folder for the images, add the bg
image directly in the html, position the header
logo and word in the center.
Honestly I'm not sure if this is supposed to be
like this. Using position attr just feels ugly.
Could it be that I'm supposed to use it on the bg
image instead? Considering opening a branch to
test this conjecture.